+++ rtSwimTracker/src/SORTtrackingBox.cpp
@@ -29,7 +29,7 @@ bool operator!= (const TrackingBox& c1, const TrackingBox& c2)
 	return(!(c1 == c2));
 }
 
-float TrackingBox::GetIOU(const TrackingBox& bb_gt)
+double TrackingBox::GetIOU(const TrackingBox& bb_gt)
 {
 	float in = (box & bb_gt.box).area();
 	float un = box.area() + bb_gt.box.area() - in;
@@ -37,5 +37,5 @@ float TrackingBox::GetIOU(const TrackingBox& bb_gt)
 	if (un < DBL_EPSILON)
 		return 0;
 
-	return (in / un);
+	return (double)(in / un);
 }
